% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_hd.R
\name{ATM_fhd}
\alias{ATM_fhd}
\title{title ash type model for f}
\usage{
ATM_fhd(Y, El, El2, sigmae2)
}
\value{
Ef is the mean of f,  Ef2 is mean of f^2
}
\description{
description use ash type model to maxmization
}
\keyword{internal}

