% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_hd.R
\name{sigmaest}
\alias{sigmaest}
\title{title estimate the variance using anova}
\usage{
sigmaest(residualsqr)
}
\value{
sigmae2_v fitted value from the anova model
}
\description{
sigma_ij = mu + a_i + b_j
}
\keyword{internal}

